+++ index.js
@@ -42,7 +42,7 @@ class Core {
 	loadDependencies() {
 		this.logger = new Logger(this.config)
 		this.generalStore = new GeneralStore(this.config)
-		this.tokenManager = new TokenManager({generalStore: this.generalStore})
+		this.tokenManager = new TokenManager(this.config, this.generalStore, this.logger)
 		this.codeGenerator = new CodeGenerator(this.config)
 	}